# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "REST APIs to CASE_EXPED table in SQL",
    "version": "1.0.0",
    "title": "API to CASE_EXPED "
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/GetTrackingInfo": {
      "get": {
        "tags": [
          "tracking"
        ],
        "summary": "Get Tracking Info by ID",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string",
            "description": "Tracking id",
            "required": true
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/AddTrackingInfo": {
      "post": {
        "tags": [
          "tracking"
        ],
        "summary": "Add tracking info",
        "description": "",
        "operationId": "addTrackingInfo",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Request object needed",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TrackingInfo"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "TrackingInfo": {
      "type": "object",
      "properties": {
        "request_id": {
          "type": "integer",
          "format": "int64"
        },
        "carrier_id": {
          "type": "string"
        },
        "driver_contact_phone": {
          "type": "string"
        },
        "completed": {
          "type": "string"
        },
        "warnings": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Tracking"
      }
    },
...
  },
...
}